+++ mac.js
@@ -1,6 +1,5 @@
 var common = require('./common')
 var fs = require('fs-extra')
-var objectAssign = require('object-assign')
 var path = require('path')
 var plist = require('plist')
 var series = require('run-series')
@@ -13,11 +12,11 @@ function rename (basePath, oldName, newName, cb) {
 function moveHelpers (frameworksPath, appName, callback) {
   series([' Helper', ' Helper EH', ' Helper NP'].map(function (suffix) {
     return function (cb) {
-      var executableBasePath = path.join(frameworksPath, 'Electron' + suffix + '.app', 'Contents', 'MacOS')
+      var executableBasePath = path.join(frameworksPath, `Electron${suffix}.app`, 'Contents', 'MacOS')
 
-      rename(executableBasePath, 'Electron' + suffix, appName + suffix, function (err) {
+      rename(executableBasePath, `Electron${suffix}`, appName + suffix, function (err) {
         if (err) return cb(err)
-        rename(frameworksPath, 'Electron' + suffix + '.app', appName + suffix + '.app', cb)
+        rename(frameworksPath, `Electron${suffix}.app`, appName + suffix + '.app', cb)
       })
     }
   }), function (err) {
@@ -32,12 +31,12 @@ function filterCFBundleIdentifier (identifier) {
 }
 
 function signOptsWarning (name) {
-  console.warn('WARNING: osx-sign.' + name + ' will be inferred from main options')
+  console.warn(`WARNING: osx-sign.${name} will be inferred from main options`)
 }
 
 function createSignOpts (properties, platform, app) {
   // use default sign opts if osx-sign is true, otherwise clone osx-sign object
-  var signOpts = properties === true ? {identity: null} : objectAssign({}, properties)
+  var signOpts = properties === true ? {identity: null} : Object.assign({}, properties)
 
   // osx-sign options are handed off to sign module, but
   // with a few additions from main options
@@ -185,7 +184,7 @@ module.exports = {
       })
 
       // Move Helper apps/executables, then top-level .app
-      var finalAppPath = path.join(tempPath, opts.name + '.app')
+      var finalAppPath = path.join(tempPath, `${opts.name}.app`)
       operations.push(function (cb) {
         moveHelpers(frameworksPath, opts.name, cb)
       }, function (cb) {